Extract //cuttlefish/files:copy from utils:files
Bug: b/536155297
