+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk/gem5_image_unpacker.cpp
@@ -20,6 +20,7 @@
 #include "android-base/file.h"
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/copy.h"
 #include "cuttlefish/host/commands/assemble_cvd/boot_image_utils.h"
 #include "cuttlefish/host/commands/assemble_cvd/flags/boot_image.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
