+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -121,6 +121,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:users",
+        "//cuttlefish/files:copy",
         "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/flag_parser",
         "//cuttlefish/host/commands/cvd/cli:command_request",
