+++ base/cvd/cuttlefish/host/libs/config/BUILD.bazel
@@ -155,6 +155,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:host_info",
+        "//cuttlefish/files:copy",
         "//cuttlefish/host/libs/config:ap_boot_flow",
         "//cuttlefish/host/libs/config:boot_flow",
         "//cuttlefish/host/libs/config:config_utils",
